Endpoint createPost organizado
