+++ semana19/projeto-labook/package.json
@@ -16,6 +16,7 @@
    "dependencies": {
       "bcryptjs": "^2.4.3",
       "cors": "^2.8.5",
+      "dayjs": "^1.10.4",
       "dotenv": "^8.2.0",
       "express": "^4.17.1",
       "jsonwebtoken": "^8.5.1",
