+++ semana19/projeto-labook/src/business/entities/post.ts
@@ -9,6 +9,6 @@ export type Post = {
     photo: string,
     description: string,
     type: POST_TYPES,
-    createdAt: Date,
+    createdAt: number,
     authorId: string
 }
