+++ semana19/projeto-labook/src/index.ts
@@ -22,37 +22,7 @@ app.use(cors())
 
 // app.post('/users/signup', )
 // app.post('/users/login', ) 
-
-app.post('/posts/create', async (req: Request, res: Response) => {
-   try {
-      let message = "Success!"
-
-      const { photo, description, type } = req.body
-
-      const token: string = req.headers.authorization as string
-
-      const tokenData: AuthenticationData = getTokenData(token)
-
-      const id: string = generateId()
-
-      await connection("labook_posts")
-         .insert({
-            id,
-            photo,
-            description,
-            type,
-            author_id: tokenData.id
-         })
-
-      res.status(201).send({ message })
-
-   } catch (error) {
-      let message = error.sqlMessage || error.message
-      res.statusCode = 400
-
-      res.send({ message })
-   }
-})
+//app.post('/posts/create',) 
 
 app.get('/posts/:id', async (req: Request, res: Response) => {
    try {
